working for some cases, but for some philos are dying when they shouldn't

+++ philo/src/libft_utils.c
@@ -12,7 +12,7 @@
 
 #include "philo.h"
 
-int	ft_atoi(char *nptr)
+long	ft_atoi(char *nptr)
 {
 	int			i;
 	int			sign;
